run_scDMV {scDMV}	R Documentation

Run scDMV

Description

Run scDMV

Usage

run_scDMV(treadn, treadx, testRegion, sample8c, sample4c, m, n)

Arguments

treadn
total reads matrix

treadx
methylation reads matrix

testRegion
the list generated after site filtering and region division

sample8c
the location of the second type sample(8cell sample)

sample4c
the location of the first type sample(4cell sample)

m
the number of columns where the first gene sequence in the file is located

n
the number of columns where the last gene sequence is located

Details

This function estimates the four parameter values of the scDMV model based on the total reads and methylation reads given by the user and then run scDMV model to get pvalues and regional level methylation difference Δ(absd) between two types of samples.

Value

pvalues(scSMVpvaluem) and regional level methylation difference Δ(absd) between two types of samples.

[Package scDMV version 0.1.0 Index]